requirements.txt
pytest==8.2.2
Appium-Python-Client==3.1.1
selenium==4.18.1
pytest-html==4.1.1
pytest-metadata==3.1.1 